% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_area.R
\name{gg_area}
\alias{gg_area}
\title{Creates a coropleth map in ggplot2}
\usage{
gg_area(
  shapefile,
  var,
  log.scale = FALSE,
  guidetitle = "Guide",
  maptitle = "Areal data"
)
}
\description{
This function allows the creation of a coropleth map with ggplot2 and sf in a easier way.
}
\keyword{areal}
\keyword{data}
