% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_intensity.R
\name{gg_intensity}
\alias{gg_intensity}
\title{Creates a intensity map for point pattern data with ggplot2.}
\usage{
gg_intensity(
  shapefile,
  data,
  x,
  y,
  crs = 29193,
  alpha = 0.5,
  title = "Intensity",
  ylab = "Y",
  xlab = "X",
  palette = "RdYlGn"
)
}
\description{
Creates a intensity map for point pattern data with ggplot2.
}
\keyword{data}
\keyword{pattern}
\keyword{point}
