% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_moran.R
\name{gg_moran}
\alias{gg_moran}
\title{Complete spatial analysis for areal data}
\usage{
gg_moran(
  shapefile,
  var,
  binary.mat = TRUE,
  maptitle = "Areal Data",
  guidetitle = "Guide",
  sign = 0.05
)
}
\description{
Complete spatial analysis for areal data
}
\keyword{areal}
\keyword{data}
