% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_moranplot.R
\name{gg_moranplot}
\alias{gg_moranplot}
\title{Creates a moran plot using ggplot2.}
\usage{
gg_moranplot(shapefile, var, binary.mat = TRUE)
}
\arguments{
\item{shapefile}{A sf object to be used on the creation of the spatial structure.}

\item{var}{A variable containing the spatial data.}

\item{binary.mat}{Logical. If TRUE then the weights natrix will be binary, if FALSE it will be weighted.}
}
\description{
desc
}
\keyword{areal}
\keyword{data,}
\keyword{moran}
\keyword{plot}
