% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pattern.R
\name{gg_pattern}
\alias{gg_pattern}
\title{Automated analysis for point pattern data with spatstat and ggplot2.}
\usage{
gg_pattern(
  shapefile,
  data,
  x,
  y,
  crs = 29193,
  fun = c("G", "F"),
  nsim = 10,
  pointsize = 1,
  pointcolor = "black",
  palette = "RdYlGn",
  alpha = 0.5,
  as.list = FALSE
)
}
\description{
Automated analysis for point pattern data with spatstat and ggplot2.
}
\keyword{data}
\keyword{pattern}
\keyword{point}
