% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_points.R
\name{gg_points}
\alias{gg_points}
\title{Creates a point pattern map with ggplot2.}
\usage{
gg_points(
  shapefile,
  data,
  x,
  y,
  crs = 29193,
  pointsize = 1,
  pointcolor = "black",
  title = "Point Pattern",
  ylab = "Y",
  xlab = "X"
)
}
\description{
Creates a point pattern map with ggplot2.
}
\keyword{areal}
\keyword{data}
