% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_area.R
\name{lf_area}
\alias{lf_area}
\title{Creates a coropleth map in leaflet.}
\usage{
lf_area(
  shapefile,
  var,
  log.scale = FALSE,
  palette = "YlOrRd",
  alpha = 0.8,
  guidetitle = "Guide"
)
}
\description{
This function allows the creation of a coropleth map with leaflet and sf in a easier way.
}
\keyword{areal}
\keyword{data}
