% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_moran.R
\name{lf_moran}
\alias{lf_moran}
\title{Complete spatial analysis for areal data}
\usage{
lf_moran(shapefile, var, binary.mat = TRUE, alpha = 0.85)
}
\description{
Complete spatial analysis for areal data
}
\keyword{areal}
\keyword{data}
