% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_points.R
\name{lf_points}
\alias{lf_points}
\title{Creates a point pattern map with leaflet.}
\usage{
lf_points(
  data,
  x,
  y,
  crs = 29193,
  pointsize = 1,
  pointcolor = "black",
  title = "Point Pattern"
)
}
\description{
Creates a point pattern map with leaflet.
}
\keyword{data}
\keyword{pattern}
\keyword{point}
