% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qspatial2-package.R
\docType{package}
\name{qspatial2-package}
\alias{qspatial2}
\alias{qspatial2-package}
\title{qspatial2: What the Package Does (One Line, Title Case)}
\description{
qspatial2 provides user friendly spatial analysis by wrapping functions from the spdep
    and spatstat packages and presenting the results in ggplot2 or leaflet.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
